# src/BackEnd/Api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "心之旅",
        "version": "1.0"
    },
    "paths": {
...
        "/api/storesys/login": {
            "post": {
                "tags": [
                    "门店管理系统-授权"
                ],
                "summary": "登录",
                "operationId": "App\\Http\\Controllers\\StoreSystem\\AccountController::Login",
                "parameters": [
                    {
                        "name": "token",
                        "in": "header",
                        "description": "token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "username",
                                    "password"
                                ],
                                "properties": {
                                    "phone": {
                                        "description": "手机号码",
                                        "type": "number",
                                        "default": "13888888888"
                                    },
                                    "code": {
                                        "description": "验证码",
                                        "type": "number",
                                        "default": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "状态码",
                                            "type": "number",
                                            "example": "200"
                                        },
                                        "msg": {
                                            "description": "提示信息",
                                            "type": "string",
                                            "example": "登录成功!"
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": ""
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "失败",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "状态码",
                                            "type": "number",
                                            "example": "500"
                                        },
                                        "msg": {
                                            "description": "提示信息",
                                            "type": "string",
                                            "example": "登录失败!"
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": "错误信息"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/storesys/verifCode": {
            "post": {
                "tags": [
                    "门店管理系统-授权"
                ],
                "summary": "获取登录验证码",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "phone"
                                ],
                                "properties": {
                                    "phone": {
                                        "description": "手机号码",
                                        "type": "number",
                                        "default": "13888888888"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "状态码",
                                            "type": "number",
                                            "example": "200"
                                        },
                                        "msg": {
                                            "description": "提示信息",
                                            "type": "string",
                                            "example": "获取验证码成功!"
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": ""
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "失败",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "状态码",
                                            "type": "number",
                                            "example": "500"
                                        },
                                        "msg": {
                                            "description": "提示信息",
                                            "type": "string",
                                            "example": "获取验证码失败!"
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": "错误信息"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "security": [
        []
    ]
}